Add files via upload
found out why the unittest controlling the number of connection was failing : the method connected return if two neurons are connected but not if they are connected more than once. thus the nb of connection calculated using this method was to low ! Now the unittest is working, but my neurons are still spiking only twice during the whole simulation.